Add files via upload
Student academic advising is very important for development of wards in different institutions but always a time-consuming effort in academic life. A Web based Academic advising system has been implemented in order to solve the issues between students and the academic advisers, by making advising more convenient, take complaint from students, evaluate and generate reports, suggest possible answers which is different from the manual way of giving student advise to an automated way. This research has analyzed the existing procedures; as utilized by many institutions and then implemented computerized solutions in order to enhance their overall advising experience. The paper presents an outline of the design and implementation of a computerized e-Academic Advising System as a web-based application. The proposed system attempts to build a model that the staff and counsel can access to follow-up the student complaints and suggestions. Also, the students can register academic complaints, seek advice and more. The e-Academic advising web-based system is developed and implemented with HTML, CSS and "PHP" as a Web framework which runs via the structured query language, "SQL" as a Database Engine.